+++ src/app/views/active-loans/active-loans.component.ts
@@ -1,13 +1,11 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Subscription } from 'rxjs';
-// App Spinner
 import { NgxSpinnerService } from 'ngx-spinner';
-// App Models
-import { Loan } from './../../models/loan.model';
-// App Services
+import { Loan, LoanType, Network } from './../../models/loan.model';
+import { LoanCurator } from './../../utils/loan-curator';
+import { ApiService } from '../../services/api.service';
+import { EventsService } from '../../services/events.service';
 import { TitleService } from '../../services/title.service';
-import { ContractsService } from './../../services/contracts.service';
-import { AvailableLoansService } from '../../services/available-loans.service';
+import { LoanTypeService } from '../../services/loan-type.service';
 
 @Component({
   selector: 'app-active-loans',
@@ -16,60 +14,91 @@ import { AvailableLoansService } from '../../services/available-loans.service';
 })
 export class ActiveLoansComponent implements OnInit, OnDestroy {
   pageId = 'active-loans';
-  available: any;
-  availableLoans = true;
   loans = [];
-
-  // subscriptions
-  subscriptionAvailable: Subscription;
+  // pagination
+  page = 0;
+  isLoading: boolean;
+  isFullScrolled: boolean;
+  isAvailableLoans = true;
 
   constructor(
     private spinner: NgxSpinnerService,
+    private apiService: ApiService,
     private titleService: TitleService,
-    private contractsService: ContractsService,
-    private availableLoansService: AvailableLoansService
+    private eventsService: EventsService,
+    private loanTypeService: LoanTypeService
   ) { }
 
   ngOnInit() {
     this.titleService.changeTitle('Activity explorer');
-    this.spinner.show(this.pageId);
     this.loadLoans();
-
-    // Available Loans service
-    this.subscriptionAvailable = this.availableLoansService.currentAvailable.subscribe(
-      available => this.available = available
-    );
   }
 
   ngOnDestroy() {
-    this.spinner.hide(this.pageId);
+    this.loading = false;
+  }
+
+  async onScroll(event: any) {
+    if (this.loading || this.isFullScrolled) {
+      return;
+    }
+
+    const { offsetHeight, scrollTop, scrollHeight } = event.target;
+    const TOLERANCE_PX = 570; // aprox card height
+    if (offsetHeight + scrollTop >= (scrollHeight - TOLERANCE_PX)) {
+      await this.loadLoans(this.page);
+    }
+  }
+
+  /**
+   * Load active loans
+   * @param page Page
+   * @return Loans
+   */
+  async loadLoans(page: number = this.page) {
+    this.loading = true;
 
     try {
-      this.subscriptionAvailable.unsubscribe();
-    } catch (e) { }
-  }
+      const diasporeLoans: Loan[] = await this.apiService.getPaginatedActiveLoans(Network.Diaspore, page);
+      const basaltLoans: Loan[] = await this.apiService.getPaginatedActiveLoans(Network.Basalt, page);
+      const loans: Loan[] = diasporeLoans.concat(basaltLoans);
+      const curatedLoans: Loan[] = LoanCurator.curateLoans(loans);
 
-  /**
-   * Update available loans number
-   */
-  upgradeAvaiblable() {
-    this.availableLoansService.updateAvailable(this.loans.length);
-  }
+      const ALLOWED_TYPES = [LoanType.UnknownWithCollateral, LoanType.FintechOriginator, LoanType.NftCollateral];
+      const filteredLoans: Loan[] = this.loanTypeService.filterLoanByType(curatedLoans, ALLOWED_TYPES);
 
-  /**
-   * Load loans
-   */
-  async loadLoans() {
-    const loans: Loan[] = await this.contractsService.getActiveLoans();
-    this.loans = loans;
+      // if there are no more loans
+      if (!loans.length) {
+        this.isFullScrolled = true;
+        this.isAvailableLoans = this.loans.length ? true : false;
+      }
 
-    this.upgradeAvaiblable();
-    this.spinner.hide(this.pageId);
+      // if there are more loans add them and continue
+      if (loans.length) {
+        this.loans = this.loans.concat(filteredLoans);
+        this.page++;
+      }
 
-    if (!this.loans.length) {
-      this.availableLoans = false;
-    } else {
-      this.availableLoans = true;
+      if (loans.length && !filteredLoans.length) {
+        await this.loadLoans();
+      }
+    } catch (err) {
+      this.eventsService.trackError(err);
+    } finally {
+      this.loading = false;
     }
+  }
+
+  private set loading(loading: boolean) {
+    this.isLoading = loading;
+    if (loading) {
+      this.spinner.show(this.pageId);
+    } else {
+      this.spinner.hide(this.pageId);
+    }
+  }
+
+  private get loading() {
+    return this.isLoading;
   }
 }